functionality
It is now able to use game logic, it will know when a player has won a
game or if it's a draw. The game can be reset by pressing 'r'. I was
having issues pulling the last commit, so i just copy pasted my stuff
instead. The main is back to what i had before Matt changed it, I did
save a copy of his edited file though.